video-transcript-pipeline/requirements.txt
mcp>=0.9.0
playwright>=1.40.0
yt-dlp>=2023.12.30
supabase>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0

